LPD-92681 Move SecretException to its own exception package
Per review feedback, give SecretException a dedicated exception package and export it, rather than keeping it beside the manager API.

Also from the same review: say "get" rather than "list" in the identifier log message, and initialize the manager at its field declaration in the test.
